[WebAuthn] CtapAuthenticator::tryDowngrade should distinguish requestData().options https://bugs.webkit.org/show_bug.cgi?id=203771 <rdar://problem/56832618>
Reviewed by Brent Fulgham.

CtapAuthenticator::tryDowngrade assumes the request is a getAssertion request. This is true
for most cases but not for Google's registration. For Google's registration, a quirk is
implemented to use U2F commands for registration if GoogleLegacyAppIdSupportExtension asks
us to do so.

No tests given there is no way to simulate google.com in our test infrastructures.

* UIProcess/WebAuthentication/fido/CtapAuthenticator.cpp:
(WebKit::CtapAuthenticator::tryDowngrade):


Identifier: 217145@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@251979 268f45cc-cd09-0410-ab3c-d52691b4dbfc