[GTK] Inspector protocol tests timing out on the bots https://bugs.webkit.org/show_bug.cgi?id=122571
Reviewed by Devin Rousso.

Mark more inspector tests as passing on GTK.

* inspector/dom/focus.html: Remove 'focus' event listener before closing dummy inspector
window. Otherwise it triggers a WebPage::setActivityState which in turn triggers focus
event on the page and on the focused element which results in two extra 'focus' lines printed.
To be clear this only avoids the extra text output that could be added to the output. Timeout
and crash problems must have been fixed before (likely by the recent changes in local inspector
client).
* platform/gtk/TestExpectations:


Identifier: 217186@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@252027 268f45cc-cd09-0410-ab3c-d52691b4dbfc