[iOS] [UIWKDocumentContext] Add the ability to request all marked text rects https://bugs.webkit.org/show_bug.cgi?id=204278 <rdar://problem/56527803>
Reviewed by Tim Horton.

Source/WebCore:

Mark a constructor as WEBCORE_EXPORT. See WebKit ChangeLog for more details.

* editing/TextIterator.h:

Source/WebKit:

Add support for a new UIWKDocumentRequest option to request character rects for each character in the marked
text range. Unless UIWKDocumentRequestRects is additionally specified (in which case we'll return rects for
every single character in the editable root anyways), this option will ensure that all characters in the
composition range (which should match -[UIWKDocumentContext markedTextRange]) will have known character rects
that can be retrieved using -[UIWKDocumentContext enumerateLayoutRects:] (or one of the other helper methods
that utilize -enumerateLayoutRects:).

Test: WebKit.DocumentEditingContextWithMarkedText

* Platform/spi/ios/UIKitSPI.h:

Add a staging declaration for the new option.

* Shared/DocumentEditingContext.h:
* UIProcess/ios/WKContentViewInteraction.mm:
(toWebDocumentRequestOptions):
* WebProcess/WebPage/ios/WebPageIOS.mm:
(WebKit::WebPage::requestDocumentEditingContext):

Refactor logic that uses CharacterIterator to find rects for each character range into a lambda function, and
use this lambda function to handle UIWKDocumentRequestRects and the new UIWKDocumentRequestMarkedTextRects. The
marked text range is relative to the start of the context before the selection.

Tools:

Add an API test to exercise the new request option.

* TestWebKitAPI/Tests/WebKitCocoa/DocumentEditingContext.mm:
(-[UIWKDocumentContext markedTextRects]):
* TestWebKitAPI/ios/UIKitSPI.h:


Identifier: 217563@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@252528 268f45cc-cd09-0410-ab3c-d52691b4dbfc