Fix getTotalLength() and getPointAtLength() for optimized rect and ellipse renderers https://bugs.webkit.org/show_bug.cgi?id=204213
Patch by Said Abou-Hallawa <sabouhallawa@apple.com> on 2019-11-18
Reviewed by Simon Fraser.

LayoutTests/imported/w3c:

* web-platform-tests/svg/types/elements/SVGGeometryElement-rect-expected.txt:

Source/WebCore:

If the renderer does not create a Path to draw the shape, we need to create
a temporary Path from the SVGElement then use it to answer the questions
of these functions and finally delete it.

* platform/graphics/Path.cpp:
(WebCore::Path::traversalStateAtLength const):
(WebCore::Path::pointAtLength const):
No need for the 'success'. It is never used.

(WebCore::Path::normalAngleAtLength const): Deleted.
This function is not used.

* platform/graphics/Path.h:
* rendering/svg/RenderSVGShape.cpp:
(WebCore::RenderSVGShape::updateShapeFromElement):
Creating the Path from the SVGELement is moved to createPath().

(WebCore::RenderSVGShape::getTotalLength const):
(WebCore::RenderSVGShape::getPointAtLength const):
Create a temporary Path if the renderer draws the shape without creating
the Path.

(WebCore::RenderSVGShape::createPath const):
* rendering/svg/RenderSVGShape.h:

* rendering/svg/SVGTextLayoutEngine.cpp:
(WebCore::SVGTextLayoutEngine::layoutTextOnLineOrPath):
* svg/SVGAnimateMotionElement.cpp:
(WebCore::SVGAnimateMotionElement::buildTransformForProgress):
Use PathTraversalState.success() to check instead of getting the same
value through an argument to traversalStateAtLength().

LayoutTests:

The following tests were added in r231955. The SVGElements in these were
defined such that a path has to be created to render any of them. Change
all the tests, except the polygon and the polyline, such that no path is
needed to render them. This will exercise the code path where we need to
create a temporary Path to answer the questions of getTotalLength() and
getPointAtLength().

* svg/dom/SVGGeometry-circle-expected.txt:
* svg/dom/SVGGeometry-circle.xhtml:
* svg/dom/SVGGeometry-ellipse-expected.txt:
* svg/dom/SVGGeometry-ellipse.xhtml:
* svg/dom/SVGGeometry-line-expected.txt:
* svg/dom/SVGGeometry-line.xhtml:
* svg/dom/SVGGeometry-polygon-expected.txt:
* svg/dom/SVGGeometry-polygon.xhtml:
* svg/dom/SVGGeometry-polyline-expected.txt:
* svg/dom/SVGGeometry-polyline.xhtml:
* svg/dom/SVGGeometry-rect-expected.txt:
* svg/dom/SVGGeometry-rect.xhtml:

Identifier: 217594@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@252563 268f45cc-cd09-0410-ab3c-d52691b4dbfc