Flex layout triggers excessive layout on height percentage descendants https://bugs.webkit.org/show_bug.cgi?id=204319 <rdar://problem/57236652>
Reviewed by Simon Fraser.

This is very similar to r252562, except in this case the layout is explicitly triggered by the flex layout logic.
The patch ensures that we don't try to lay out percent height descendants with out-of-flow ancestors (see r252562 or webkit.org/b/204255 for more info).
(Unfortunately this is not testable but the subsequent repaint fix will include a test which would fail if we regressed this.)

* rendering/RenderFlexibleBox.cpp:
(WebCore::RenderFlexibleBox::hasPercentHeightDescendants const):
(WebCore::RenderFlexibleBox::layoutAndPlaceChildren):
* rendering/RenderFlexibleBox.h:


Identifier: 217614@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@252620 268f45cc-cd09-0410-ab3c-d52691b4dbfc