REGRESSION(r249428): [GStreamer] VP9 video rendered green https://bugs.webkit.org/show_bug.cgi?id=201422
Patch by Thibault Saunier <tsaunier@igalia.com> on 2019-11-19
Reviewed by Philippe Normand.

Avoid forcing a video conversion in software we end up and make
it happen in GL with a sensibly simpler pipeline, this is possible as
`glcolorconvert` properly setups the converted GL buffers.

Without that patch we are getting "random" not negotiated issues with
MediaStream sources related to the `GLMemory` caps feature during
renegotiation between the software `videoconvert` and our sink.

* platform/graphics/gstreamer/MediaPlayerPrivateGStreamerBase.cpp:
(WebCore::MediaPlayerPrivateGStreamerBase::createVideoSinkGL):

Identifier: 217618@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@252624 268f45cc-cd09-0410-ab3c-d52691b4dbfc