[Clipboard API] Add support for Clipboard.readText() https://bugs.webkit.org/show_bug.cgi?id=204310 <rdar://problem/57292754>
Reviewed by Ryosuke Niwa.

LayoutTests/imported/w3c:

Rebaseline a couple of imported web platform tests.

* web-platform-tests/clipboard-apis/async-navigator-clipboard-basics.https-expected.txt:
* web-platform-tests/clipboard-apis/async-write-text-read-text-manual.https-expected.txt:

Source/WebCore:

Implements readText; see below for more detail.

Tests: editing/async-clipboard/clipboard-do-not-read-text-from-platform-if-text-changes.html
       editing/async-clipboard/clipboard-read-text-from-platform.html
       editing/async-clipboard/clipboard-read-text-same-origin.html
       editing/async-clipboard/clipboard-read-text.html

* Modules/async-clipboard/Clipboard.cpp:
(WebCore::Clipboard::readText):

Implement the method. This works similarly to Clipboard::read, but immediately reads text data for the first
clipboard item with data for "text/plain", instead of exposing a list of clipboard items.

* dom/DataTransfer.cpp:
(WebCore::DataTransfer::commitToPasteboard):

If the custom pasteboard data object is empty, then don't bother trying to write it to the platform pasteboard.
This avoids hitting an assertion in WebDragClient::beginDrag that checks to make sure there's no data on the
pasteboard right before beginning a drag in the UI process which was getting hit because we'd otherwise end up
writing just the pasteboard data origin; it also has an added bonus of avoiding an unnecessary sync IPC message.

* platform/mac/PlatformPasteboardMac.mm:
(WebCore::PlatformPasteboard::write):

Write the custom data type ("com.apple.WebKit.custom-pasteboard-data") when writing PasteboardCustomData to the
pasteboard on macOS, even when the custom pasteboard data only contains an origin. This allows DOM paste
requests to automatically accept when exchanging text data between the same origin via the async clipboard API.

LayoutTests:

Add several new layout tests.

* editing/async-clipboard/clipboard-do-not-read-text-from-platform-if-text-changes-expected.txt: Added.
* editing/async-clipboard/clipboard-do-not-read-text-from-platform-if-text-changes.html: Added.

Add a test to verify that if the clipboard changes content in the middle of a call to clipboard.readText, we
will reject the readText() promise and avoid exposing any text to the page.

* editing/async-clipboard/clipboard-read-text-expected.txt: Added.
* editing/async-clipboard/clipboard-read-text-from-platform-expected.txt: Added.
* editing/async-clipboard/clipboard-read-text-from-platform.html: Added.

Add a test to verify that we display DOM paste UI when reading text that was written to the pasteboard directly
via platform API.

* editing/async-clipboard/clipboard-read-text-same-origin-expected.txt: Added.
* editing/async-clipboard/clipboard-read-text-same-origin.html: Added.

Add a test to verify that we allow the page to access same origin text data on the pasteboard using readText,
during a user gesture.

* editing/async-clipboard/clipboard-read-text.html: Added.

Add a basic test to verify that readText works when the page writes text to the clipboard using DataTransfer and
document.execCommand.

* platform/gtk/TestExpectations:
* platform/mac-wk1/TestExpectations:
* platform/win/TestExpectations:
* platform/wpe/TestExpectations:


Identifier: 217621@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@252627 268f45cc-cd09-0410-ab3c-d52691b4dbfc