Fix Tools/Scripts/run-javascriptcore-tests after r252490 https://bugs.webkit.org/show_bug.cgi?id=204452
Rubber-stamped by Philippe Normand

Regression introduced in:

  results.webkit.org: Report JSC tests to the results database
  https://bugs.webkit.org/show_bug.cgi?id=204091
  <rdar://problem/49778900>

WebKitGTK and WPE build environment is defined by the dependencies
policy document [1] as a Debian Stable. For Debian the `uname` binary
is installed in the `/bin/` directory and this directory is included
in the `$PATH` for any regular or system user. This fix removes the
path for the `uname` command in the run-javascriptcore-tests script.

[1] https://trac.webkit.org/wiki/WebKitGTK/DependenciesPolicy

* Scripts/run-javascriptcore-tests:
(configurationForUpload):


Identifier: 217715@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@252734 268f45cc-cd09-0410-ab3c-d52691b4dbfc