[iOS] Web process crashes when attempting to request document context in a textarea https://bugs.webkit.org/show_bug.cgi?id=204455 <rdar://problem/57383975>
Reviewed by Tim Horton.

Source/WebCore:

Make `WebCore::rangesOverlap` robust in the case where the ranges' container nodes are in different tree scopes.
Currently, `compareBoundaryPoints` returns an exception in this scenario, which `rangesOverlap` does not expect.
As a result, we crash when releasing the return value after boundary point comparison. To fix this, just upgrade
the `ownerDocument` check in `rangesOverlap` to check for tree scopes instead, via the
`areNodesConnectedInSameTreeScope` helper function.

Test: WebKit.DocumentEditingContextSpatialRequestInTextField

* dom/Range.cpp:
(WebCore::rangesOverlap):

Tools:

Add a test to verify that we don't crash (and are still able to retrieve character rects).

* TestWebKitAPI/Tests/WebKitCocoa/DocumentEditingContext.mm:
(-[UIWKDocumentContext markedTextRects]):
(-[UIWKDocumentContext textRects]):
* TestWebKitAPI/ios/UIKitSPI.h:


Identifier: 217724@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@252743 268f45cc-cd09-0410-ab3c-d52691b4dbfc