Remove unneeded code that annotated DOMHTMLElement as conforming to UITextInputTraits protocol https://bugs.webkit.org/show_bug.cgi?id=204464
Reviewed by Wenson Hsieh.

DOMHTMLElement never really conformed to the UITextInputTraits protocol. It was annotated as doing so
in order to register with the runtime as part of advertising that it implemented two UITextInputTraits_Private
messages in r222487: -acceptsAutofilledLoginCredentials and -representingPageURL. These messages were
subsequently removed in r222991 (over years ago). So, we do not need to continue advertising UITextInputTraits
conformance.

* DOM/DOMHTMLInputElement.h:

Identifier: 217727@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@252746 268f45cc-cd09-0410-ab3c-d52691b4dbfc